Move Blockchain and Block into blockchain.zig module
use [N]u8 inplace of []const u8 to make serialization easy.
This is a temporary fix till pointer swizzling and unswizzling is implemented in the serializer

remove allocator in Block and use stack for all allocations
since allocator is removed deinit fn in Block isn't need anymore so remove it

Add a db field to Blockchain for storing blocks in the db.
Currently block db storing and serialization isn't very robust and
there are some latent bugs to be fixed. More testing is required to
improve robustness